jarvis/requirements.txt
faster-whisper
sounddevice
numpy
scipy
requests
pywin32
PyQt6
edge-tts
pygame-ce
chromadb
pyautogui
pygetwindow
beautifulsoup4
duckduckgo-search
psutil
plyer
spotipy
pyperclip
wolframalpha
piper-tts
soundfile
python-dotenv
playwright            # For Autonomous Web Agent
playwright-stealth    # To prevent bot detection
watchdog              # For Shadow Learner (FileSystem monitoring)
fastapi               # Web Dashboard API
uvicorn               # ASGI Server
websockets            # Real-time HUD syncing
opencv-python         # For Presence Sensing
pytesseract           # For Screen Sentinel OCR
